+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod project;
 pub mod resolution;
 pub mod server;
 pub mod sessions;
+pub mod updater;
 pub mod util;
 pub mod vfs;
 pub mod workspace;