+++ src/updater.rs
@@ -1,0 +1,6 @@
+use anyhow::Result;
+
+pub fn check_for_updates() -> Result<()> {
+	// TODO
+	Ok(())
+}