+++ CHANGELOG.md
@@ -1,17 +1,17 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), that adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Fix Argon installer to work with GitHub Actions
 
 ## [2.0.0-pre1] - 2024-03-18
 
 ### Added
 
--   Brand new Argon CLI, written in Rust
+- Brand new Argon CLI, written in Rust
 
 [Unreleased]: https://github.com/argon-rbx/argon/compare/2.0.0-pre1...HEAD
-
 [2.0.0-pre1]: https://github.com/argon-rbx/argon/compare/3057ca895492519fc29e7ab0bd8bdebc86d3e53c...2.0.0-pre1